+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Release Notes
 * **IMPROVED:** Updates to pages are now processed and applied to pages
 * **IMPROVED:** Packet X/25 is now decoded and exported
 * **FIXED:** Subpages are now exported in the right order
+* **FIXED:** Fastext link numbers are no longer corrupted on magazine 8
 
 1.1
 ---